requirements.txt
Flask==2.3.3
PyJWT==2.8.1
requests==2.31.0
cryptography==41.0.7